initial v6 push

+++ secda_tools/secda_profiler/profiler.cc
@@ -15,8 +15,7 @@ ClockCycles::ClockCycles(string _name, bool _resetOnSave) {
 }
 
 int SignalTrack::readCount() {
-  if (resetOnSave)
-    value = 0;
+  if (resetOnSave) value = 0;
   return value;
 }
 
@@ -27,8 +26,7 @@ SignalTrack::SignalTrack(string _name) {
 }
 
 void SignalTrack::increment(int val) {
-  if (values.size() < val + 1)
-    values.resize(val + 1);
+  if (values.size() < val + 1) values.resize(val + 1);
   values[val]++;
 }
 
@@ -40,8 +38,7 @@ SignalTrack::SignalTrack(string _name, bool _resetOnSave) {
 }
 
 int ClockCycles::readCount() {
-  if (resetOnSave)
-    value = 0;
+  if (resetOnSave) value = 0;
   return value;
 }
 
@@ -78,24 +75,21 @@ void Profile::saveProfile(vector<Metric *> captured_metrics) {
           reinterpret_cast<ClockCycles *>(captured_metrics[i]);
       ClockCycles temp(capped_metric->name);
       temp.value = capped_metric->value;
-      if (capped_metric->resetOnSave)
-        capped_metric->value = 0;
+      if (capped_metric->resetOnSave) capped_metric->value = 0;
       newRecord.push_back(temp);
     } else if (captured_metrics[i]->type == TDataCount) {
       DataCount *capped_metric =
           reinterpret_cast<DataCount *>(captured_metrics[i]);
       DataCount temp(capped_metric->name);
       temp.value = capped_metric->value;
-      if (capped_metric->resetOnSave)
-        capped_metric->value = 0;
+      if (capped_metric->resetOnSave) capped_metric->value = 0;
       newRecord.push_back(temp);
     } else if (captured_metrics[i]->type == TBufferSpace) {
       BufferSpace *capped_metric =
           reinterpret_cast<BufferSpace *>(captured_metrics[i]);
       BufferSpace temp(capped_metric->name, capped_metric->total);
       temp.value = capped_metric->value;
-      if (capped_metric->resetOnSave)
-        capped_metric->value = 0;
+      if (capped_metric->resetOnSave) capped_metric->value = 0;
       newRecord.push_back(temp);
     } else if (captured_metrics[i]->type == TDataCountArray) {
       DataCountArray *capped_metric =
@@ -103,8 +97,7 @@ void Profile::saveProfile(vector<Metric *> captured_metrics) {
       for (int l = 0; l < capped_metric->value; l++) {
         DataCount temp(capped_metric->name + to_string(l));
         temp.value = capped_metric->array[l];
-        if (capped_metric->resetOnSave)
-          capped_metric->array[l] = 0;
+        if (capped_metric->resetOnSave) capped_metric->array[l] = 0;
         newRecord.push_back(temp);
       }
     } else if (captured_metrics[i]->type == TSignalTrack) {
@@ -116,11 +109,9 @@ void Profile::saveProfile(vector<Metric *> captured_metrics) {
           temp.value = capped_metric->values[l];
           newRecord.push_back(temp);
         }
-        if (capped_metric->resetOnSave)
-          capped_metric->values[l] = 0;
+        if (capped_metric->resetOnSave) capped_metric->values[l] = 0;
       }
-      if (capped_metric->resetOnSave)
-        capped_metric->value = 0;
+      if (capped_metric->resetOnSave) capped_metric->value = 0;
     }
   }
   records.push_back(newRecord);
@@ -136,24 +127,21 @@ void Profile::saveBlank(vector<Metric *> captured_metrics) {
           reinterpret_cast<ClockCycles *>(captured_metrics[i]);
       ClockCycles temp(capped_metric->name);
       temp.value = -1;
-      if (capped_metric->resetOnSave)
-        capped_metric->value = 0;
+      if (capped_metric->resetOnSave) capped_metric->value = 0;
       newRecord.push_back(temp);
     } else if (captured_metrics[i]->type == TDataCount) {
       DataCount *capped_metric =
           reinterpret_cast<DataCount *>(captured_metrics[i]);
       DataCount temp(capped_metric->name);
       temp.value = -1;
-      if (capped_metric->resetOnSave)
-        capped_metric->value = 0;
+      if (capped_metric->resetOnSave) capped_metric->value = 0;
       newRecord.push_back(temp);
     } else if (captured_metrics[i]->type == TBufferSpace) {
       BufferSpace *capped_metric =
           reinterpret_cast<BufferSpace *>(captured_metrics[i]);
       BufferSpace temp(capped_metric->name, capped_metric->total);
       temp.value = -1;
-      if (capped_metric->resetOnSave)
-        capped_metric->value = 0;
+      if (capped_metric->resetOnSave) capped_metric->value = 0;
       newRecord.push_back(temp);
     } else if (captured_metrics[i]->type == TDataCountArray) {
       DataCountArray *capped_metric =
@@ -161,8 +149,7 @@ void Profile::saveBlank(vector<Metric *> captured_metrics) {
       for (int l = 0; l < capped_metric->value; l++) {
         DataCount temp(capped_metric->name + to_string(l));
         temp.value = -1;
-        if (capped_metric->resetOnSave)
-          capped_metric->array[l] = 0;
+        if (capped_metric->resetOnSave) capped_metric->array[l] = 0;
         newRecord.push_back(temp);
       }
     }
@@ -173,7 +160,7 @@ void Profile::saveBlank(vector<Metric *> captured_metrics) {
 void Profile::addMetric(Metric m) { model_record.push_back(m); }
 
 void Profile::updateMetric(Metric m) {
-  for (int i = 0; i < model_record.size(); i++) {
+  for (int i = 0; i < (int)model_record.size(); i++) {
     if (m.name == model_record[i].name) {
       model_record[i] = m;
     }
@@ -181,7 +168,7 @@ void Profile::updateMetric(Metric m) {
 }
 
 void Profile::incrementMetric(string name, int value) {
-  for (int i = 0; i < model_record.size(); i++) {
+  for (int i = 0; i < (int)model_record.size(); i++) {
     if (name == model_record[i].name) {
       model_record[i].value += value;
     }
@@ -191,25 +178,20 @@ void Profile::incrementMetric(string name, int value) {
 void Profile::saveCSVRecords(string filename) {
 #ifdef SYSC
 
-  if (records.size() == 0)
-    return;
+  if (records.size() == 0) return;
   ofstream per_sim_file;
   per_sim_file.open(filename + ".csv");
 
   for (int i = 0; i < base_metrics.size(); i++) {
     per_sim_file << base_metrics[i].name;
-    if (i + 1 != base_metrics.size())
-      per_sim_file << ",";
-    else
-      per_sim_file << endl;
+    if (i + 1 != base_metrics.size()) per_sim_file << ",";
+    else per_sim_file << endl;
   }
 
   for (int i = 0; i < records[0].size(); i++) {
     per_sim_file << records[0][i].name;
-    if (i + 1 != records[0].size())
-      per_sim_file << ",";
-    else
-      per_sim_file << endl;
+    if (i + 1 != records[0].size()) per_sim_file << ",";
+    else per_sim_file << endl;
   }
   std::vector<std::string> metric_names;
   for (int r = 0; r < records.size(); r++) {
@@ -227,19 +209,15 @@ void Profile::saveCSVRecords(string filename) {
       for (int i = 0; i < records[r].size(); i++) {
         if (records[r][i].name == metric_names[m]) {
           per_sim_file << records[r][i].value;
-          if (m + 1 != metric_names.size())
-            per_sim_file << ",";
-          else
-            per_sim_file << endl;
+          if (m + 1 != metric_names.size()) per_sim_file << ",";
+          else per_sim_file << endl;
           found = true;
         }
       }
       if (!found) {
         per_sim_file << "0";
-        if (m + 1 != metric_names.size())
-          per_sim_file << ",";
-        else
-          per_sim_file << endl;
+        if (m + 1 != metric_names.size()) per_sim_file << ",";
+        else per_sim_file << endl;
       }
     }
   }
@@ -247,15 +225,15 @@ void Profile::saveCSVRecords(string filename) {
 
   // ofstream per_model_file;
   // per_model_file.open(filename + "_model.csv");
-  // for (int i = 0; i < model_record.size(); i++) {
+  // for (int i = 0; i < (int) model_record.size(); i++) {
   //   per_model_file << model_record[i].name;
-  //   if (i + 1 != model_record.size()) per_model_file << ",";
+  //   if (i + 1 != (int) model_record.size()) per_model_file << ",";
   //   else per_model_file << endl;
   // }
 
-  // for (int m = 0; m < model_record.size(); m++) {
+  // for (int m = 0; m < (int) model_record.size(); m++) {
   //   per_model_file << model_record[m].value;
-  //   if (m + 1 != model_record.size()) per_model_file << ", ";
+  //   if (m + 1 != (int) model_record.size()) per_model_file << ", ";
   //   else per_model_file << endl;
   // }
   // per_model_file.close();
